refactor: Update table.jsx
