+++ src/pages/Trainee/components/Table/Style.js
@@ -1,0 +1,43 @@
+import TableRow from '@material-ui/core/TableRow';
+import { makeStyles, withStyles } from '@material-ui/core/styles';
+
+const StyledTableRow = withStyles((theme) => ({
+  root: {
+    '&:nth-of-type(odd)': {
+      backgroundColor: theme.palette.action.hover,
+    },
+  },
+}))(TableRow);
+
+const useStyles = makeStyles((theme) => ({
+  container: {
+    marginTop: theme.spacing(2),
+    marginBottom: theme.spacing(2),
+  },
+  table: {
+    minWidth: 650,
+    border: 'solid',
+    borderWidth: 'thin',
+    borderColor: 'lightGrey',
+  },
+  style: {
+    cursor: 'pointer',
+  },
+  color: {
+    color: 'grey',
+  },
+  visuallyHidden: {
+    border: 0,
+    clip: 'rect(0 0 0 0)',
+    height: 1,
+    margin: -1,
+    overflow: 'hidden',
+    padding: 0,
+    position: 'absolute',
+    top: 20,
+    width: 1,
+  },
+
+}));
+
+export { StyledTableRow, useStyles };
